[cpp_project] GAMPSLimit: use the calculated epsilon for encoding the ratio and the base columns

+++ cpp_project/include/coders/constant/apca/apca_window.h
@@ -9,19 +9,19 @@
 class APCAWindow: public Window {
 
 private:
-    double min;
-    double max;
+    int min;
+    int max;
 #if !MASK_MODE
     bool nan_window;
 #endif
 
 public:
-    APCAWindow(int window_size_, int error_threshold_);
+    APCAWindow(int window_size_, double error_threshold_);
     bool conditionHolds(std::string x);
     bool isFull();
     bool isEmpty();
     void addFirstValue(std::string x);
-    bool updateConstantValue(double new_min, double new_max);
+    bool updateConstantValue(int new_min, int new_max);
 };
 
 #endif //CPP_PROJECT_APCA_WINDOW_H
