+++ cpp_project/include/coders/utils/window.h
@@ -8,7 +8,7 @@ class Window {
 
 protected:
     int window_size;
-    int error_threshold;
+    double error_threshold;
 
 public:
     int window_size_bit_length;
@@ -17,7 +17,7 @@ public:
 
     Window(){};
 
-    Window(int window_size_, int error_threshold_){
+    Window(int window_size_, double error_threshold_){
         window_size = window_size_;
         error_threshold = error_threshold_;
         window_size_bit_length = MathUtils::windowSizeBitLength(window_size);
